search_api/pyproject.toml
[project]
name = "search_api"
version = "0.1.0"
description = "Judgment Search API"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "qdrant-client>=1.7.0",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.0.0",
    "neo4j>=5.0.0",
    "networkx>=3.0.0",
    # 高度なNLP技術（軽量版）
    "spacy[ja]>=3.7.0",
    # オントロジー構築
    "rdflib>=7.0.0",
    "owlready2>=0.45",
    # テキスト処理
    "janome>=0.4.2",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ruff>=0.12.8",
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "bandit>=1.8.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "bandit>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disable_error_code = ["no-any-return"]
